Replace double with big decimal

+++ app/src/main/java/com/bobowiec/revolut_app/data/model/CurrentRates.kt
@@ -1,3 +1,5 @@
 package com.bobowiec.revolut_app.data.model
 
-data class CurrentRates(val base: String, val date: String, val rates: Map<String, Double>)
+import java.math.BigDecimal
+
+data class CurrentRates(val base: String, val date: String, val rates: Map<String, BigDecimal>)
